added all users tab to explore
